% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\docType{data}
\name{demo_biseq_DMR}
\alias{demo_biseq_DMR}
\title{DMR BiSeq Demo Dataset}
\description{
The BiSeq dataset for demo purpose. The data are dummy data. It includes 9 columns:

The dummy output for BiSeq illustrating purpose. It is dummy.
}
\details{
- seqnames: Chromosome

 - start: The positions of the start sites of the corresponding region

 - end: The positions of the end sites of the corresponding region

 - strand: Strand

 - median.p

 - median.meth.group1

 - median.meth.group2

 - median.meth.diff

- seqnames

- start

- end

- width

- strand

- median.p

- median.meth.group1

- median.meth.group2

- median.meth.diff

Notice that there are "NaN" within the feature columns.

Please check the vignette "metevalue" for details.
}
\keyword{metevalue}
