% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\docType{data}
\name{demo_biseq_methyrate}
\alias{demo_biseq_methyrate}
\title{BiSeq Methyrate Demo Dataset}
\description{
The methyrate for BiSeq illustrating purpose. It is dummy.
}
\details{
The data includes 12 columns.

- chr: string Chromosome

- pos: int Position

- g1~g2: methylation rate data in groups, repeat 5 times.
Notice that there are "NaN" within the feature columns.

Please check the vignette "metevalue" for details.
}
\keyword{metevalue}
