% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\docType{data}
\name{demo_desq_out}
\alias{demo_desq_out}
\title{DESeq Output Dataset}
\description{
The output dummy data for "RNA" meythod illustrating purpose.
}
\details{
The data includes 10 columns.

- treated1fb:
- treated2fb:
- treated3fb:
- untreated1fb:
- untreated2fb:

This data contains 8166 rows and 7 columns.

Please check the vignette "metevalue" for details.
}
\keyword{The}
\keyword{a}
\keyword{data}
\keyword{data:}
\keyword{is}
\keyword{metevalue}
\keyword{simulation}
