% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\docType{data}
\name{demo_DMRfinder_DMRs}
\alias{demo_DMRfinder_DMRs}
\title{DMRfinder Output Demo Dataset}
\description{
The output dummy dataset for DMRfinder illustrating purpose.
}
\details{
The data includes 6 columns.

- chr: string Chromosome

- pos: int Position

- g1~g2: methylation rate data in groups, repeat 2 times.
Notice that there are "NaN" within the feature columns.

Please check the vignette "metevalue" for details.
}
\keyword{metevalue}
