% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\docType{data}
\name{demo_DMRfinder_rate_combine}
\alias{demo_DMRfinder_rate_combine}
\title{DMRfinder Methyrate Demo Dataset}
\description{
The methyrate for BiSeq illustrating purpose. It is dummy.
}
\details{
The data includes 6 columns.

- chr: string Chromosome

- pos: int Position

- g1~g2: methylation rate data in groups.
  
Notice that there are "NaN" within the feature columns.

Please check the vignette "metevalue" for details.
}
\keyword{metevalue}
