% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\docType{data}
\name{demo_methylkit_met_all}
\alias{demo_methylkit_met_all}
\title{Methyrate output dataset from methylKit}
\description{
The methyrate dataset samples "myCpG" data from the methylKit (a bioconductor package) for illustrating purpose.
}
\details{
The data includes 7 columns:

 - chr:   Chromosome

 - start: The positions of the start sites of the corresponding region

 - end: The positions of the end sites of the corresponding region

 - strand: Strand

 - pvalue: The adjusted p-value based on BH method in MWU-test

 - qvalue: cutoff for qvalue of differential methylation statistic

 - methyl.diff: The difference between the group means of methylation level

Please check the vignette "metevalue" for details.
}
\references{
Akalin, Altuna, et al. "methylKit: a comprehensive R package for the analysis of genome-wide DNA methylation profiles." Genome biology 13.10 (2012): 1-9. \doi{10.1186/gb-2012-13-10-r87}
}
\keyword{metevalue}
