% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\docType{data}
\name{demo_methylkit_methyrate}
\alias{demo_methylkit_methyrate}
\title{Methyrate Dataset}
\description{
The methyrate dataset samples "myCpG" data from the methylKit (a bioconductor package) for illustrating purpose.
}
\details{
The data includes 6 columns.

- chr: string Chromosome

- pos: int Position

- g1~g2: methylation rate data in groups (4 columns)

Please check the vignette "metevalue" for details.
}
\references{
Akalin, Altuna, et al. "methylKit: a comprehensive R package for the analysis of genome-wide DNA methylation profiles." Genome biology 13.10 (2012): 1-9. \doi{10.1186/gb-2012-13-10-r87}
}
\keyword{metevalue}
