% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\docType{data}
\name{demo_metilene_out}
\alias{demo_metilene_out}
\title{Metilene Demo Output Dataset}
\description{
The output dummy data for "metilene" meythod illustrating purpose.
}
\details{
The data includes 10 columns.

- V1: string Chromosome

- V2: The positions of the start sites of the corresponding region

- V3: The positions of the end sites of the corresponding region

- V4- V10: data value.

Please check the vignette "metevalue" for details.
}
\keyword{metevalue}
