% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_data_chk.R
\name{evalue.biseq.chk}
\alias{evalue.biseq.chk}
\title{Check the BiSeq data format}
\usage{
evalue.biseq.chk(
  input_filename_a,
  input_filename_b,
  sep = "\\t",
  bheader = FALSE
)
}
\arguments{
\item{input_filename_a}{metilene input file path. This file is a sep (e.g. TAB) separated file with two key columns and several value columns in pairs:
For exampe:

chrom    pos    g1    g1    g1    g1    g1    g1    g1    g1    g2    g2    g2    g2    g2    g2    g2    g2

chrom and pos are keys;
g1 g1 g2 g2 must be stored in pairs.}

\item{input_filename_b}{metilene input file path. This file should stored as a sep(e.g. TAB) separated file with two key columns and several value columns:
The columns are (in order):

    - chr:   Chromosome

    - start: The position of the start site of the corresponding region

    - end: The position of the end site of the corresponding region

    - range: The range of the corresponding region

    - strand: Strand

    - median.p:  The median of p-values in the corresponding region

    - median.meth.group1 : The median of methylation level for the corresponding segment of group 1

    - median.meth.group2 : The median of methylation level for the corresponding segment of group 2

    - median.meth.diff:  The median of the difference between the methylation level}

\item{sep}{separator, default is the TAB key.}

\item{bheader}{a logical value indicating whether the input_filename_b file contains the names of the variables as its first line. By default, bheader = FALSE.}
}
\value{
list(file_a, file_b, file_a_b) returns a list with three pr-handled data.frames corresponding to the input_filename_a, input_filename_b file and a A JOIN B file.
}
\description{
Check the BiSeq data format
}
\examples{
\dontrun{
s = evalue.metilene(input_filename_a = "../metilene/metilene_8.input",
input_filename_b = "../metilene/metilene_8.out")
## > str(s)
## data.frame':	723 obs. of  11 variables:
## $ chr        : chr  "chr21" "chr21" "chr21" "chr21" ...
## $ start      : int  9437432 9708982 9825467 9825514 9825794 9825882 9826220 9827335 9926563 9963396 ...
## $ end        : int  9437540 9709189 9825508 9825788 9825876 9826191 9826387 9827356 9927097 9963753 ...
## $ q-value    : num  2.49e-25 4.62e-29 6.00e-02 3.40e-01 2.82e-07 ...
## $ methyl.diff: num  0.611 0.476 -0.274 -0.164 -0.261 ...
## $ CpGs       : int  26 28 12 26 26 31 13 10 73 10 ...
## $ p         : num  3.86e-14 4.34e-14 2.60e-07 2.55e-05 1.23e-11 ...
## $ p2        : num  2.23e-29 4.13e-33 5.37e-06 3.04e-05 2.52e-11 ...
## $ m1         : num  0.737 0.589 0.298 0.374 0.353 ...
## $ m2         : num  0.126 0.113 0.573 0.538 0.615 ...
## $ e_value    : num  8.50e+40 2.71e+38 7.20e+05 1.89e+06 5.36e+12 ...
}
}
