% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_data_DMRfinder.R
\name{evalue.DMRfinder}
\alias{evalue.DMRfinder}
\title{Evalue of the DMRfinder data format}
\usage{
evalue.DMRfinder(
  methyrate,
  DMRfinder.output,
  adjust.methods = "BH",
  sep = "\\t",
  bheader = FALSE
)
}
\arguments{
\item{methyrate}{is the methyrate file.}

\item{DMRfinder.output}{is the output file of DMRfinder.}

\item{adjust.methods}{is the adjust methods of e-value. It can be 'bonferroni', 'hochberg', 'holm', 'hommel', 'BH', 'BY'}

\item{sep}{seperator, default is the TAB key.}

\item{bheader}{a logical value indicating whether the DMRfinder.output file contains the names of the variables as its first line. By default, bheader = FALSE.}
}
\description{
Perform the Evaluation for the BiSeq data.
}
