% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_data_methylKit.R
\name{evalue.methylKit}
\alias{evalue.methylKit}
\title{Evalue of the methylKit data format}
\usage{
evalue.methylKit(
  methyrate,
  methylKit.output,
  adjust.methods = "BH",
  sep = "\\t",
  bheader = FALSE
)
}
\arguments{
\item{sep}{seperator, default is the TAB key.}

\item{bheader}{a logical value indicating whether the input_filename_b file contains the names of the variables as its first line. By default, bheader = FALSE.}

\item{methylKit.output'}{is the output file of methylKit}

\item{outputfile'}{is the output data with e-value of each region}

\item{adjust.methods'}{is the adjust methods of e-value. It can be 'bonferroni', 'hochberg', 'holm', 'hommel', 'BH', 'BY'}
}
\description{
Perform the Evaluation for the BiSeq data.
}
