% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_data_chk.R
\name{evalue_buildin_sql}
\alias{evalue_buildin_sql}
\title{Buildin data process function}
\usage{
evalue_buildin_sql(a, b, method = "metilene")
}
\arguments{
\item{a}{filename}

\item{b}{filenmae}

\item{method}{"metilene" or "biseq" or "methylKit"}
}
\description{
Buildin data process function
}
