% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_data_chk.R
\name{evalue_buildin_var_fmt_nm}
\alias{evalue_buildin_var_fmt_nm}
\title{Buildin check file format function}
\usage{
evalue_buildin_var_fmt_nm(a, b, method = "metilene")
}
\arguments{
\item{a}{filename}

\item{b}{filenmae}

\item{method}{"metilene" or "biseq" or "methylKit"}
}
\description{
Buildin check file format function
}
