% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\docType{data}
\name{met_all}
\alias{met_all}
\title{Methyrate output dataset from methylKit}
\description{
The methyrate dataset samples "myCpG" data from the methylKit (a bioconductor package) for illustrating purpose.
}
\details{
The data includes 7 columns:

 - chr:   Chromosome

 - start: The positions of the start sites of the corresponding region

 - end: The positions of the end sites of the corresponding region

 - strand: Strand

 - pvalue: The adjusted p-value based on BH method in MWU-test

 - qvalue: cutoff for qvalue of differential methylation statistic

 - methyl.diff: The difference between the group means of methylation level

Please check the vignette "metevalue" for details.
}
\references{
\url{https://www.bioconductor.org/packages/release/bioc/html/methylKit.html}
}
\keyword{metevalue}
