% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\docType{data}
\name{methyrate}
\alias{methyrate}
\title{Methyrate Dataset}
\description{
The methyrate dataset samples "myCpG" data from the methylKit (a bioconductor package) for illustrating purpose.
}
\details{
The data includes 6 columns.

- chr: string Chromosome

- pos: int Position

- g1~g2: metilene data in pairs (4 columns)

Please check the vignette "metevalue" for details.
}
\references{
\url{https://www.bioconductor.org/packages/release/bioc/html/methylKit.html}
}
\keyword{metevalue}
