% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\name{varevalue.metilene}
\alias{varevalue.metilene}
\title{Evalue of the Metilene data}
\usage{
varevalue.metilene(a, b, a_b, adjust.methods = "BH")
}
\arguments{
\item{a}{A data.frame object, the columns should be (in order):

chrom    pos    g1    g1    g1    g1    g1    g1    g1    g1    g2    g2    g2    g2    g2    g2    g2    g2

i.e two key columns (chrom, pos) with several value columns in pairs.}

\item{b}{A data.frame object stores the data, the columns are (in order):

    - chr:   Chromosome

    - start: The positions of the start sites of the corresponding region

    - end: The positions of the end sites of the corresponding region

    - q-value: The adjusted p-value based on BH method in MWU-test

    - methyl.diff: The difference between the group means of methylation level

    - CpGs:  The number of CpG sites within the corresponding region

    - p : p-value based on MWU-test

    - p2: p-value based on 2D KS-test

    - m1:  The absolute mean methylation level for the corresponding segment of group 1

    - m2:  The absolute mean methylation level for the corresponding segment of group 2}

\item{a_b}{A data.frame object of a join b with particular data clean processes. Check the function [evalue.methylKit.chk()] for more details.}

\item{adjust.methods}{is the adjust methods of e-value. It can be 'bonferroni', 'hochberg', 'holm', 'hommel', 'BH', 'BY'. The default value is 'BH'.}
}
\value{
a dataframe, the columns are (in order):

    - chr:   Chromosome

    - start: The positions of the start sites of the corresponding region

    - end: The positions of the end sites of the corresponding region

    - q-value: The adjusted p-value based on BH method in MWU-test

    - methyl.diff: The difference between the group means of methylation level

    - CpGs:  The number of CpG sites within the corresponding region

    - p : p-value based on MWU-test

    - p2: p-value based on 2D KS-test

    - m1:  The absolute mean methylation level for the corresponding segment of group 1

    - m2:  The absolute mean methylation level for the corresponding segment of group 2

    - e_value: The e-value of the corresponding region
}
\description{
Perform the Evaluation for the Metilene data. The data file could be pre-handled by the evalue.metilene.chk function.
}
\examples{
\dontrun{
s = evalue.metilene(input_filename_a = "metilene.input",
input_filename_b = "metilene.out")
## > str(s)
## data.frame':	723 obs. of  11 variables:
## $ chr        : chr  "chr21" "chr21" "chr21" "chr21" ...
## $ start      : int  9437432 9708982 9825467 ...
## $ end        : int  9437540 9709189 9825508 ...
## $ q-value    : num  2.49e-25 4.62e-29 6.00e-02 3.40e-01 2.82e-07 ...
## $ methyl.diff: num  0.611 0.476 -0.274 -0.164 -0.261 ...
## $ CpGs       : int  26 28 12 26 26 31 13 10 73 10 ...
## $ p         : num  3.86e-14 4.34e-14 2.60e-07 2.55e-05 1.23e-11 ...
## $ p2        : num  2.23e-29 4.13e-33 5.37e-06 3.04e-05 2.52e-11 ...
## $ m1         : num  0.737 0.589 0.298 0.374 0.353 ...
## $ m2         : num  0.126 0.113 0.573 0.538 0.615 ...
## $ e_value    : num  8.50e+40 2.71e+38 7.20e+05 1.89e+06 5.36e+12 ...
}
}
