% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\name{varevalue.single_general}
\alias{varevalue.single_general}
\title{A general method to calculate the e-value for other DNA methylation tools not described above. The input data is the DNA methylation rates using the same format with Metilene.}
\usage{
varevalue.single_general(
  methyrate,
  group1_name = "g1",
  group2_name = "g2",
  chr,
  start,
  end
)
}
\arguments{
\item{methyrate}{data.frame: A data.frame object of methylation rates, the columns should be(name of groups can be self-defined)

chr    pos    group1_name group1_name ... group1_name group2_name group2_name}

\item{group1_name}{charactor: The name of the first group. For example, "treated" in the above example.}

\item{group2_name}{charactor: The name of the second group. For example, "untreated" in the above example.}

\item{chr}{charactor: The Chromosome name. Typically, it is a string like "chr21" and so on.}

\item{start}{integer:  The position of the start site of the corresponding region}

\item{end}{integer: The position of the end site of the corresponding region}
}
\value{
evalue
}
\description{
The data file could be pre-handled by the metevalue.[types].chk function.
The  Chromosome name, start and end sits shoule be specified.
}
\examples{
#data("demo_metilene_input")
#varevalue.single_general(demo_metilene_input, chr = "chr21", start = 9437432, end = 9437540)
# [1] 2.626126e+43

#### Compare to `metevalue.metilene`  ####
data(demo_metilene_out)
#example_tempfiles = tempfile(c("metilene_input", "metilene_out"))
#tempdir()
#write.table(demo_metilene_input, file=example_tempfiles[1],
#      row.names=FALSE, col.names=TRUE, quote=FALSE, sep='\t')
#write.table (demo_metilene_out, file=example_tempfiles[2],
#      sep ="\t", row.names =FALSE, col.names =TRUE, quote =FALSE)
#result = metevalue.metilene(example_tempfiles[1], example_tempfiles[2],
#      bheader = TRUE)
# result[with(result, chr == 'chr21' & start == '9437432' & end == '9437540'), ncol(result)]
# [1] 2.626126e+43
}
