% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\docType{data}
\name{demo_desq_out}
\alias{demo_desq_out}
\title{DESeq Output Dataset}
\description{
The output dummy data for "RNA" meythod illustrating purpose.
}
\details{
The data includes 10 columns.

- treated1fb:

- treated2fb:

- treated3fb:

- untreated1fb:

- untreated2fb:

- untreated3fb:
 
- untreated4fb:

This data contains 8166 rows and 7 columns.

Please check the vignette "metevalue" for details.
}
\examples{
# library("pasilla")
# pasCts <- system.file("extdata",
#                       "pasilla_gene_counts.tsv",
#                       package="pasilla", mustWork=TRUE)
# pasAnno <- system.file("extdata",
#                        "pasilla_sample_annotation.csv",
#                        package="pasilla", mustWork=TRUE)
# cts <- as.matrix(read.csv(pasCts,sep="\t",row.names="gene_id"))
# coldata <- read.csv(pasAnno, row.names=1)
# coldata <- coldata[,c("condition","type")]
# coldata$condition <- factor(coldata$condition)
# coldata$type <- factor(coldata$type)
# 
# library("DESeq2")
# colnames(cts)=paste0(colnames(cts),'fb')
# cts = cts[,rownames(coldata)]
# dds <- DESeqDataSetFromMatrix(countData = cts,
#                               colData = coldata,
#                               design = ~ condition)
# dds <- DESeq(dds)
# 
# 
# dat <- t(t(cts)/(dds$sizeFactor)) 
# dat.out <- dat[rowSums(dat >5)>=0.8*ncol(dat),]
# 
# demo_desq_out <- log(dat.out)
}
\keyword{The}
\keyword{a}
\keyword{data}
\keyword{data:}
\keyword{is}
\keyword{metevalue}
\keyword{simulation}
