% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_data_chk.R
\name{evalue_buildin_sql}
\alias{evalue_buildin_sql}
\title{Build-in data process function}
\usage{
evalue_buildin_sql(a, b, method = "metilene")
}
\arguments{
\item{a}{data frame of the methylation rate}

\item{b}{data frame of output data corresponding to the "method" option}

\item{method}{"metilene" or "biseq", "DMRfinder" or "methylKit"}
}
\value{
a data frame combines data frame a and b corresponding to the "method" option
}
\description{
Build-in data process function
}
\examples{
\donttest{
data("demo_metilene_out")
data("demo_metilene_input")
result = evalue_buildin_var_fmt_nm(demo_metilene_input,
                                   demo_metilene_out, method="metilene")
result_sql = evalue_buildin_sql(result$a, result$b, method="metilene")
}
}
