% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\name{metevalue.RNA_general}
\alias{metevalue.RNA_general}
\title{A general method to calculate the e-value for RNA data.}
\usage{
metevalue.RNA_general(a_b, group1_name, group2_name)
}
\arguments{
\item{a_b}{data.frame: A data.frame object of a join b with particular data clean processes. Check the function [evalue.methylKit.chk()] for more details. 
Notice, this data frame only contains the value in pairs, for example:


     treated1fb treated2fb untreated1fb untreated2fb

TAG1   4.449648   4.750104     4.392285     4.497514

TAG2   8.241116   8.302852     8.318125     8.488796

...


Row names (TAG1 and TAG2 in the above example) is also suggested.}

\item{group1_name}{charactor: The name of the first group. For example, "treated" in the above example.}

\item{group2_name}{charactor: The name of the second group. For example, "untreated" in the above example.}
}
\value{
evalue
}
\description{
Perform the Evaluation for the RNA data.
}
\examples{
data("demo_desq_out")
evalue = evalue_RNA(demo_desq_out, 'treated','untreated')
}
