% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue_main.R
\name{metevalue.RNA_general}
\alias{metevalue.RNA_general}
\title{A general method to calculate the e-value for RNA-seq data.}
\usage{
metevalue.RNA_general(rna, group1_name, group2_name)
}
\arguments{
\item{rna}{data.frame: A data.frame object of RNAseq data. For example:


     treated1fb treated2fb untreated1fb untreated2fb

TAG1   4.449648   4.750104     4.392285     4.497514

TAG2   8.241116   8.302852     8.318125     8.488796

...


Row names (TAG1 and TAG2 in the above example) is also suggested.}

\item{group1_name}{charactor: The name of the first group. For example, "treated" in the example.}

\item{group2_name}{charactor: The name of the second group. For example, "untreated" in the example.}
}
\value{
evalue
}
\description{
A general method to calculate the e-value for RNA-seq data.
}
\examples{
data("demo_desq_out")
evalue = metevalue.RNA_general(demo_desq_out, 'treated','untreated')
}
